+++ src/richie/apps/courses/urls.py
@@ -1,15 +1,17 @@
 """
 API routes exposed by our Courses app.
 """
+from django.urls import re_path
+
 from rest_framework import routers
 
-from .api import CourseRunsViewSet
+from .api import CourseRunsViewSet, course_runs_sync
 
-# Instantiate a router
 ROUTER = routers.SimpleRouter()
 
-# Define our app's routes with the router
-ROUTER.register(r"course-runs", CourseRunsViewSet, "course-runs")
+# Define our app's API routes with the router
+ROUTER.register("course-runs", CourseRunsViewSet, "course_runs")
 
-# Use the standard name for our urlpatterns so urls.py can import it effortlessly
-urlpatterns = ROUTER.urls
+urlpatterns = ROUTER.urls + [
+    re_path("course-runs-sync/?$", course_runs_sync, name="course_run_sync"),
+]