+++ tests/apps/courses/lms/test_lms_select.py
@@ -4,24 +4,22 @@ from django.test.utils import override_settings
 
 from richie.apps.courses.lms import LMSHandler
 from richie.apps.courses.lms.base import BaseLMSBackend
-from richie.apps.courses.lms.edx import TokenEdXLMSBackend
+from richie.apps.courses.lms.edx import EdXLMSBackend
 
 
 class LMSSelectTestCase(TestCase):
     """Test suite for the LMS select function."""
 
     @override_settings(
-        LMS_BACKENDS=[
+        RICHIE_LMS_BACKENDS=[
             {
                 "COURSE_REGEX": r"^.*/moocs/(?P<course_id>.*)",
-                "SELECTOR_REGEX": r".*example.com.*",
-                "BACKEND": "richie.apps.courses.lms.edx.TokenEdXLMSBackend",
+                "BACKEND": "richie.apps.courses.lms.edx.EdXLMSBackend",
                 "BASE_URL": "https://www.example.com",
             },
             {
                 "COURSE_REGEX": r"^.*/courses/(?P<course_id>.*)",
-                "SELECTOR_REGEX": r".*edx.org.*",
-                "BACKEND": "richie.apps.courses.lms.edx.BaseLMSBackend",
+                "BACKEND": "richie.apps.courses.lms.base.BaseLMSBackend",
                 "BASE_URL": "https://edx.org",
             },
         ]
@@ -33,7 +31,7 @@ class LMSSelectTestCase(TestCase):
         any course id can be extract from the resource_link.
         """
         backend = LMSHandler.select_lms("https://www.example.com/moocs/123")
-        self.assertEqual(type(backend), TokenEdXLMSBackend)
+        self.assertEqual(type(backend), EdXLMSBackend)
         self.assertEqual(backend.configuration["BASE_URL"], "https://www.example.com")
 
         backend = LMSHandler.select_lms("https://edx.org/courses/123")